Move battle pet account durability behind typed port
