+++ crates/wow-database/src/lib.rs
@@ -33,6 +33,7 @@
 //! - [`SqlTransaction`]: Batch of statements executed atomically
 //! - Statement enums: [`LoginStatements`], [`WorldStatements`], [`CharStatements`], [`HotfixStatements`]
 
+pub mod battle_pet_account_adapter;
 pub mod database;
 pub mod error;
 pub mod gameobject_use_template_adapter;
@@ -60,6 +61,7 @@ pub mod updater;
 pub mod void_storage_adapter;
 
 // Re-export primary types at crate root for convenience.
+pub use battle_pet_account_adapter::LoginBattlePetPersistenceLikeCpp;
 pub use database::{
     Database, build_connection_string, build_connection_string_with_ssl_like_cpp,
     escape_string_like_cpp, warn_about_sync_queries_enabled_like_cpp,
